refactor(components): :recycle: Move Rect/Circle Serializers into Serialize.hpp
Serializer<math::Rect> and Serializer<math::Circle> lived in Collider.hpp
even though nothing about them is Collider-specific -- they're just the
TOML (de)serializers for the two shape types, referenced by any component
that needs to write a shape out (Collider today, Animation's frame rects
next). Moving them to Serialize.hpp, next to the generic Serializer<T>
template, makes that reusability explicit instead of forcing an unrelated
component to include Collider.hpp just to reuse its shape serializers.

TOMLTableView also gains GetTable(path, index) alongside the existing
GetTable(path) -- equivalent to GetTable(path + "[" + index + "]"), for
callers walking a known element count out of an array-of-tables without
building that bracket-indexed string themselves.
